backend/worker/requirements.txt
pika==1.3.2
minio==7.2.10
Pillow=11.0.0
python-dotenv==1.0.1